+++ src/backends/rest.ts
@@ -1,14 +1,20 @@
-import {Headers, Http, RequestMethod, RequestOptions, URLSearchParams} from "@angular/http";
-import {Observable} from "rxjs/Observable";
-import {IDSBackend, IDSBackendProvider} from "./interface";
-import {DSJsonParser} from "../parsers/json";
-import {Inject, Injectable, InjectionToken, Optional} from "@angular/core";
-import {DSJsonRenderer} from "../renderers/json";
+import { Headers, Http, RequestMethod, RequestOptions, URLSearchParams } from "@angular/http";
+import { Observable } from "rxjs/Observable";
+import { IDSBackend, IDSBackendConfig, IDSBackendProvider } from "./interface";
+import { Inject, Injectable, InjectionToken, Optional } from "@angular/core";
 import "rxjs/add/operator/map";
 import "rxjs/observable/of";
 import "rxjs/operator/mergeMap";
-import {capitalize, isString} from "lodash";
+import { capitalize, defaults, isString } from "lodash";
+import { DSConfig } from "../configuration";
+import { IDSResponseParser, IDSResponseParserProvider } from "../parsers/interface";
+import { IDSRequestRenderer, IDSRequestRendererClass } from "../renderers/interface";
+import { IDSAuthentication, IDSAuthenticationClass } from "../authentication/interface";
+import { IDSAuthenticationProvider, IDSRequestRendererProvider, IDSResponseParserClass } from "../ngdataservice";
 
+/**
+ * Rest detail : parts of an url.
+ */
 export interface IDSRestIdentifier {
     path: string;
     query?: { [index: string]: string };
@@ -21,34 +27,51 @@ export interface IDSRestBackendConfig {
     port: string;
     scheme: string;
     url?: string;
-    headers?: { [index: string]: string }
+    headers?: { [index: string]: string };
 }
 
 export let REST_BACKEND_CONFIG = new InjectionToken<IDSRestBackendConfig>("backend.rest.config");
+export let REST_BACKEND_PARSER = new InjectionToken<IDSResponseParser>("backend.rest.parser");
+export let REST_BACKEND_RENDERER = new InjectionToken<IDSRequestRenderer>("backend.rest.renderer");
+export let REST_BACKEND_AUTHENTICATION = new InjectionToken<IDSAuthentication>("backend.rest.authentication");
 
-// FIXME: AOT: wait for https://github.com/angular/angular/issues/12631
-/* export class DSRestBackendConfig implements IDSRestBackendConfig {
- public host: string;
- public port: string;
- public scheme: string;
- public url: string;
- public headers: {[index: string]: string}
- } */
+@Injectable()
+export class DSRestBackendSetup implements IDSBackendConfig {
+    public constructor(@Inject(REST_BACKEND_AUTHENTICATION) public authentication: IDSAuthentication,
+                       @Inject(REST_BACKEND_PARSER) public parser: IDSResponseParser,
+                       @Inject(REST_BACKEND_RENDERER) public renderer: IDSRequestRenderer) {
+    }
+
+}
 
 
 /**
  * Makes data requests to a REST API.
- * FIXME: include authentication in backend ?
  * TODO: add custom actions.
  */
 @Injectable()
-export class DSRestBackend implements IDSBackend {
+export class DSRestBackend extends DSConfig<IDSBackendConfig> implements IDSBackend {
+    public authentication_class: IDSAuthenticationClass;
+    public authentication_provider: IDSAuthenticationProvider;
+    public authentication_config: any;
+    public parser_class: IDSResponseParserClass;
+    public parser_provider: IDSResponseParserProvider;
+    public parser_config: any;
+    public renderer_class: IDSRequestRendererClass;
+    public renderer_provider: IDSRequestRendererProvider;
+    public renderer_config: any;
     private _defaultHeaders: { [index: string]: string } = {};
+    /* tslint:disable:no-unused-variable */
+    private _renderer: IDSRequestRenderer;
+    private _parser: IDSResponseParser;
+    private _authentication: IDSAuthentication;
+    /* tslint:enable */
 
     constructor(private _http: Http,
-                protected _parser: DSJsonParser,
-                protected _renderer: DSJsonRenderer,
+                @Optional() protected setup: DSRestBackendSetup,
                 @Inject(REST_BACKEND_CONFIG) protected _config: IDSRestBackendConfig) {
+        super();
+        this.setup = defaults(this.setup || {}, setup);
         if (this._config.headers) {
             this._defaultHeaders = this._config.headers;
         }
@@ -56,63 +79,63 @@ export class DSRestBackend implements IDSBackend {
 
     public retrieve(identifier: IDSRestIdentifier, params: any = {}): Observable<any> {
         let options = this.getRequestOptions(identifier);
-        options = this._renderer.prepare(options);
+        options = this.renderer.prepare(options);
         return this._http
             .get(this.getRequestUrl(identifier), options)
             .map((response) => {
-                return this._parser.parse(response);
+                return this.parser.parse(response);
             });
     }
 
     public list(identifier: IDSRestIdentifier, params: any = {}): Observable<any> {
         let options = this.getRequestOptions(identifier);
-        options = this._renderer.prepare(options);
+        options = this.renderer.prepare(options);
         console.log("Calling list http");
         return this._http
             .get(this.getRequestUrl(identifier), options)
             .map((response) => {
-                return this._parser.parse(response);
+                return this.parser.parse(response);
             });
     }
 
     public create(identifier: IDSRestIdentifier, values: any, params: any = {}): Observable<any> {
         let options = this.getRequestOptions(identifier);
-        options = this._renderer.prepare(options);
+        options = this.renderer.prepare(options);
         return this._http
-            .post(this.getRequestUrl(identifier), this._renderer.render(values), options)
+            .post(this.getRequestUrl(identifier), this.renderer.render(values), options)
             .map((response) => {
-                return this._parser.parse(response);
+                return this.parser.parse(response);
             });
     }
 
     public update(identifier: IDSRestIdentifier, values: any, params: any = {}): Observable<any> {
         let options = this.getRequestOptions(identifier);
-        options = this._renderer.prepare(options);
+        options = this.renderer.prepare(options);
         return this._http
-            .put(this.getRequestUrl(identifier), this._renderer.render(values), options)
+            .put(this.getRequestUrl(identifier), this.renderer.render(values), options)
             .map((response) => {
-                return this._parser.parse(response);
+                return this.parser.parse(response);
             });
     }
 
     public partial_update(identifier: IDSRestIdentifier, values: any, params: any = {}): Observable<any> {
         let options = this.getRequestOptions(identifier);
-        options = this._renderer.prepare(options);
+        options = this.renderer.prepare(options);
         return this._http
-            .patch(this.getRequestUrl(identifier), this._renderer.render(values), options)
+            .patch(this.getRequestUrl(identifier), this.renderer.render(values), options)
             .map((response) => {
-                return this._parser.parse(response);
+                return this.parser.parse(response);
             });
     }
 
     public destroy(identifier: IDSRestIdentifier, params: any = {}): Observable<any> {
         let options = this.getRequestOptions(identifier);
-        options = this._renderer.prepare(options);
+        options = this.renderer.prepare(options);
         return this._http
             .delete(this.getRequestUrl(identifier), options)
             .map((response) => {
                 if (response.status !== 204) {
-                    return this._parser.parse(response);
+                    return this.parser.parse(response);
                 } else {
                     return null;
                 }
@@ -123,18 +146,18 @@ export class DSRestBackend implements IDSBackend {
         // Params : string => url
         // Params : object => {url, body}
         let options = this.getRequestOptions(identifier);
-        options = this._renderer.prepare(options);
+        options = this.renderer.prepare(options);
         options.method = RequestMethod[capitalize(action)];
         if (isString(params)) {
             options.url = this.getRequestUrl(identifier) + <string>params || "";
         } else {
             options.url = this.getRequestUrl(identifier) + params.url || "";
-            options.body = this._renderer.render(params.body);
+            options.body = this.renderer.render(params.body);
         }
         return this._http.request(options.url, options)
             .map((response) => {
                 console.log("ACTION : ", response);
-                return this._parser.parse(response);
+                return this.parser.parse(response);
             });
     }
 
@@ -147,7 +170,7 @@ export class DSRestBackend implements IDSBackend {
 
     /**
      * Compute request headers.
-     * @param identifier Rest identifier
+     * @param identifier Rest detail
      * @returns {Headers} Headers object
      */
     public getRequestHeaders(identifier: IDSRestIdentifier): Headers {
@@ -155,12 +178,13 @@ export class DSRestBackend implements IDSBackend {
         for (let h of Object.keys(this._defaultHeaders)) {
             headers.set(h, this._defaultHeaders[h]);
         }
+        this.authentication.setAuthHeaders(headers);
         return headers;
     }
 
     /**
      * Compute request query string (search).
-     * @param identifier Rest identifier
+     * @param identifier Rest detail
      * @returns {URLSearchParams} Url search params option
      */
     public getSearchParams(identifier: IDSRestIdentifier): URLSearchParams {
@@ -187,7 +211,7 @@ export class DSRestBackend implements IDSBackend {
     }
 
     /**
-     * Compute request url using config and identifier.
+     * Compute request url using config and detail.
      * TODO: check trailing slashes
      * SEE: add basePath in detailed config ?
      * SEE: add current host as default host ?
@@ -204,14 +228,52 @@ export class DSRestBackend implements IDSBackend {
                 identifier.path;
         }
     }
+
+
+    public get authentication(): IDSAuthentication {
+        return <IDSAuthentication>this.get_service("authentication", this.get_authentication_config());
+    }
+
+    public get_authentication_config(): any {
+        return this.get_service_config("authentication");
+    }
+
+    public set authentication(au: IDSAuthentication) {
+        this._authentication = au;
+    }
+
+    public get renderer(): IDSRequestRenderer {
+        return <IDSRequestRenderer>this.get_service("renderer", this.get_renderer_config());
+    }
+
+    public get_renderer_config(): any {
+        return this.get_service_config("renderer");
+    }
+
+    public set renderer(au: IDSRequestRenderer) {
+        this._renderer = au;
+    }
+
+    public get parser(): IDSResponseParser {
+        return <IDSResponseParser>this.get_service("parser", this.get_parser_config());
+    }
+
+    public get_parser_config(): any {
+        return this.get_service_config("parser");
+    }
+
+    public set parser(au: IDSResponseParser) {
+        this._parser = au;
+    }
+
+
 }
 
 
 @Injectable()
 export class DSRestBackendProvider implements IDSBackendProvider {
     constructor(protected _http: Http,
-                protected _parser: DSJsonParser,
-                protected _renderer: DSJsonRenderer,
+                protected setup: DSRestBackendSetup,
                 @Optional() @Inject(REST_BACKEND_CONFIG) protected _config: IDSRestBackendConfig) {
 
     }
@@ -219,8 +281,7 @@ export class DSRestBackendProvider implements IDSBackendProvider {
     public provide(params: IDSRestBackendConfig): IDSBackend {
         return new DSRestBackend(
             this._http,
-            this._parser,
-            this._renderer,
+            this.setup,
             params || this._config
         );
     }