+++ src/collection/interface.ts
@@ -35,14 +35,6 @@ export interface IDSCollectionSetup {
     adapter_class?: IDSAdapterClass;
     adapter_provider?: IDSAdapterProvider;
     adapter_config?: any;
-    backend?: IDSBackend;
-    backend_class?: IDSBackendClass;
-    backend_provider?: IDSBackendProvider;
-    backend_config?: any;
-    authentication?: IDSAuthentication;
-    authentication_class?: IDSAuthenticationClass;
-    authentication_provider?: IDSAuthenticationProvider;
-    authentication_config?: any;
     serializer?: IDSSerializer;
     serializer_class?: IDSSerializerClass;
     serializer_provider?: IDSSerializerProvider;
@@ -64,6 +56,10 @@ export interface IDSCollectionSetup {
     sorter_class?: IDSSorterClass;
     sorter_provider?: IDSSorterProvider;
     sorter_config?: any;
+    backend?: IDSBackend;
+    backend_class?: IDSBackendClass;
+    backend_provider?: IDSBackendProvider;
+    backend_config?: any;
 }
 
 export interface IDSCollectionCreateParams {
@@ -86,6 +82,7 @@ export type IDSCollectionRefreshParams = any;
 export interface IDSCollectionActionFullParams {
     url: string;
     body: any;
+    context?: any;
 }
 export type IDSCollectionActionParams = IDSCollectionActionFullParams | string;
 
@@ -96,10 +93,10 @@ export interface IDSCollection<T extends IDSModel> {
     datasources: IDSRegister;
     queryset: IDSQueryset<T>;
     adapter: IDSAdapter;
-    backend: IDSBackend;
     serializer: IDSSerializer;
     persistence: IDSPersistence;
-    authentication: IDSAuthentication;
+    backend: IDSBackend;
+
 
     save(instance: T, params?: IDSCollectionSaveParams): Observable<T>;
     update(instance: T, fields: string[], params?: IDSCollectionUpdateParams): Observable<T>;
@@ -118,7 +115,6 @@ export interface IDSCollectionConstructor<T extends IDSModel> {
 export const COLLECTION_SETUP_NAMES: string[] = [
     "adapter", "adapter_class", "adapter_provider", "adapter_config",
     "backend", "backend_class", "backend_provider", "backend_config",
-    "authentication", "authentication_class", "authentication_provider", "authentication_config",
     "serializer", "serializer_class", "serializer_provider", "serializer_config",
     "persistence", "persistence_class", "persistence_provider", "persistence_config",
     "paginator", "paginator_class", "paginator_provider", "paginator_config",