+++ src/collection/restcollection.ts
@@ -1,21 +1,20 @@
-import {DSCollection} from "./collection";
-import {IDSModel} from "../model/interface";
-import {DSRestUrlAdapterProvider} from "../adapters/resturl";
-import {DSRestBackend} from "../backends/rest";
-import {DSDefaultSerializer} from "../serializers/default";
-import {DSMemoryPersistence} from "../persistence/memory";
-import {DSTokenAuthentication} from "../authentication/tokenauth";
-import {Injectable, Optional} from "@angular/core";
-import {IDSCollectionSetup, IDSCollectionContext} from "./interface";
-import {DSDummyFilterProvider} from "../filters/dummy";
-import {DSDummySorterProvider} from "../sorters/dummy";
-import {DSDummyPaginatorProvider} from "../paginators/dummy";
+import { DSCollection } from "./collection";
+import { IDSModel } from "../model/interface";
+import { DSRestUrlAdapterProvider } from "../adapters/resturl";
+import { DSRestBackend } from "../backends/rest";
+import { DSDefaultSerializer } from "../serializers/default";
+import { DSMemoryPersistence } from "../persistence/memory";
+import { DSRestAuthentication } from "../authentication/tokenauth";
+import { Injectable, Optional } from "@angular/core";
+import { IDSCollectionContext, IDSCollectionSetup } from "./interface";
+import { DSDummyFilterProvider } from "../filters/dummy";
+import { DSDummySorterProvider } from "../sorters/dummy";
+import { DSDummyPaginatorProvider } from "../paginators/dummy";
 
 @Injectable()
 export class DSRestCollectionSetup implements IDSCollectionSetup {
     public constructor(public adapter_provider: DSRestUrlAdapterProvider,
                        public backend: DSRestBackend,
-                       public authentication: DSTokenAuthentication,
                        public serializer: DSDefaultSerializer,
                        public persistence: DSMemoryPersistence,
                        public paginator_provider: DSDummyPaginatorProvider,